Reorganize more ansible-test code. (#74611)
* Split out shell command.
* Relocate ansible-test integration code.
