+++ changelogs/fragments/ansible-test-refactor-integration.yml
@@ -1,0 +1,2 @@
+minor_changes:
+  - ansible-test - Reorganize integration test implementation by command.
