+++ changelogs/fragments/ansible-test-shell-command.yml
@@ -1,0 +1,2 @@
+minor_changes:
+  - ansible-test - Split out shell command implementation.
