+++ test/lib/ansible_test/_internal/cli.py
@@ -22,6 +22,7 @@ from .util import (
     read_lines_without_comments,
     MAXFD,
     ANSIBLE_TEST_DATA_ROOT,
+    SUPPORTED_PYTHON_VERSIONS,
 )
 
 from .delegation import (
@@ -30,15 +31,26 @@ from .delegation import (
 )
 
 from .executor import (
-    command_posix_integration,
-    command_network_integration,
-    command_windows_integration,
-    command_shell,
-    SUPPORTED_PYTHON_VERSIONS,
     ApplicationWarning,
     Delegate,
     generate_pip_install,
     configure_pypi_proxy,
+)
+
+from .commands.integration.posix import (
+    command_posix_integration,
+)
+
+from .commands.integration.network import (
+    command_network_integration,
+)
+
+from .commands.integration.windows import (
+    command_windows_integration,
+)
+
+from .commands.shell import (
+    command_shell,
 )
 
 from .config import (
